requirements.txt
asreview==1.5
datasets==2.14.7
numpy==1.26.2
pandas==2.1.3
pyalex==0.13
python-dotenv==1.0.0
pyvespa==0.39.0
requests==2.31.0
sentence_transformers==2.2.2
synergy-dataset==1.2